% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_catchments_layer}
\alias{get_catchments_layer}
\title{Get sea-draining catchments layer}
\usage{
get_catchments_layer(api_key)
}
\arguments{
\item{api_key}{Set your API key for MfE}
}
\value{
A SF object
}
\description{
Get sea-draining catchments layer
}
