% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_coastline_layer}
\alias{get_coastline_layer}
\title{Get coastlines layer}
\usage{
get_coastline_layer(api_key, layer_type = c("polygon", "line"))
}
\arguments{
\item{api_key}{Set your API key for LINZ}

\item{layer_type}{The type of layer to get. Options are 'polygon' or 'line'}
}
\value{
A Sf object
}
\description{
Get coastlines layer
}
