% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_layer_as_sf}
\alias{get_layer_as_sf}
\title{Get a spatial layer and return it as an sf object}
\usage{
get_layer_as_sf(api_key, agency, id, custom_url = NULL)
}
\arguments{
\item{api_key}{The API key for the agency}

\item{agency}{The agency from which to get the layer. Options are 'mfe', 'linz', or 'custom'}

\item{id}{The ID of the layer to get}

\item{custom_url}{The custom URL for the agency. Required if agency is 'custom'}
}
\value{
An sf object
}
\description{
Get a spatial layer and return it as an sf object
}
\examples{
# Get a layer from LINZ
result_sf <- get_layer_as_sf(api_key, agency = "linz", id = "50088")
}
