% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_regions_layer}
\alias{get_regions_layer}
\title{Get regional boundaries}
\usage{
get_regions_layer(api_key)
}
\arguments{
\item{api_key}{Set your API key for Stats NZ}
}
\value{
A Sf object
}
\description{
Get regional boundaries
}
