% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_table_as_tibble}
\alias{get_table_as_tibble}
\title{Get a table and return it as a tibble}
\usage{
get_table_as_tibble(api_key, agency, id, custom_url = NULL)
}
\arguments{
\item{api_key}{The API key for the agency}

\item{agency}{The agency from which to get the table. Options are 'mfe', 'linz', 'stats', or 'custom'}

\item{id}{The ID of the table to get}

\item{custom_url}{The custom URL for the agency. Required if agency is 'custom'}
}
\value{
A tibble
}
\description{
Get a table and return it as a tibble
}
\examples{
# Get a table from the Ministry for the Environment
result_tibble <- get_table_as_tibble(api_key, agency = "mfe", id = "53606")
}
